Abstracts away cart context into custom hook
